# packages/api-client/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/venues": {
      "get": {
        "operationId": "ExploreController_venues",
        "parameters": [
          {
            "name": "lat",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number",
              "minimum": -90,
              "maximum": 90
            }
          },
          {
            "name": "lng",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number",
              "minimum": -180,
              "maximum": 180
            }
          },
          {
            "name": "radiusKm",
            "required": false,
            "in": "query",
            "schema": {
              "default": 10,
              "type": "number",
              "minimum": 0,
              "exclusiveMinimum": true,
              "maximum": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "slug": {
                        "type": "string"
                      },
                      "name": {
                        "type": "string"
                      },
                      "type": {
                        "type": "string",
                        "enum": [
                          "GAME_BAR",
                          "TCG_SHOP",
                          "LUDOTHEQUE",
                          "ASSOCIATION",
                          "OTHER"
                        ]
                      },
                      "address": {
                        "type": "string"
                      },
                      "latitude": {
                        "type": "number"
                      },
                      "longitude": {
                        "type": "number"
                      },
                      "isPartner": {
                        "type": "boolean"
                      },
                      "kwatroPerk": {
                        "nullable": true,
                        "type": "string"
                      },
                      "distanceMeters": {
                        "type": "integer",
                        "minimum": -9007199254740991,
                        "maximum": 9007199254740991
                      },
                      "openNow": {
                        "nullable": true,
                        "type": "boolean"
                      },
                      "closesAtMinute": {
                        "nullable": true,
                        "type": "integer",
                        "minimum": -9007199254740991,
                        "maximum": 9007199254740991
                      },
                      "upcomingEventCount": {
                        "type": "integer",
                        "minimum": -9007199254740991,
                        "maximum": 9007199254740991
                      }
                    },
                    "required": [
                      "id",
                      "slug",
                      "name",
                      "type",
                      "address",
                      "latitude",
                      "longitude",
                      "isPartner",
                      "kwatroPerk",
                      "distanceMeters",
                      "openNow",
                      "closesAtMinute",
                      "upcomingEventCount"
                    ]
                  }
                }
              }
            }
          }
        },
        "tags": [
          "explore"
        ]
      }
    },
...
    "/events/{id}": {
      "get": {
        "operationId": "EventsController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "type": {
                      "type": "string",
                      "enum": [
                        "GAME_NIGHT",
                        "INITIATION",
                        "TOURNAMENT",
                        "PRERELEASE",
                        "THEMED"
                      ]
                    },
                    "description": {
                      "nullable": true,
                      "type": "string"
                    },
                    "startsAt": {
                      "type": "string",
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                    },
                    "endsAt": {
                      "nullable": true,
                      "type": "string",
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                    },
                    "cancelledAt": {
                      "nullable": true,
                      "type": "string",
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                    },
                    "priceCents": {
                      "nullable": true,
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    "capacity": {
                      "nullable": true,
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    "minAge": {
                      "nullable": true,
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    "registrationMode": {
                      "type": "string",
                      "enum": [
                        "NONE",
                        "IN_APP",
                        "EXTERNAL"
                      ]
                    },
                    "externalUrl": {
                      "nullable": true,
                      "type": "string"
                    },
                    "registeredCount": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    "games": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "slug": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "slug",
                          "name"
                        ]
                      }
                    },
                    "venue": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "slug": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "address": {
                          "type": "string"
                        },
                        "isPartner": {
                          "type": "boolean"
                        }
                      },
                      "required": [
                        "id",
                        "slug",
                        "name",
                        "address",
                        "isPartner"
                      ]
                    },
                    "myRegistration": {
                      "nullable": true,
                      "type": "string",
                      "enum": [
                        "REGISTERED",
                        "WAITLISTED"
                      ]
                    }
                  },
                  "required": [
                    "id",
                    "title",
                    "type",
                    "description",
                    "startsAt",
                    "endsAt",
                    "cancelledAt",
                    "priceCents",
                    "capacity",
                    "minAge",
                    "registrationMode",
                    "externalUrl",
                    "registeredCount",
                    "games",
                    "venue",
                    "myRegistration"
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "events"
        ]
      }
    },
...
    "/admin/venues": {
      "get": {
        "operationId": "AdminCatalogController_venues",
        "parameters": [
          {
            "name": "q",
            "required": false,
            "in": "query",
            "schema": {
              "default": "",
              "type": "string",
              "maxLength": 100
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "PUBLISHED"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "slug": {
                        "type": "string"
                      },
                      "name": {
                        "type": "string"
                      },
                      "type": {
                        "type": "string",
                        "enum": [
                          "GAME_BAR",
                          "TCG_SHOP",
                          "LUDOTHEQUE",
                          "ASSOCIATION",
                          "OTHER"
                        ]
                      },
                      "status": {
                        "type": "string",
                        "enum": [
                          "PENDING",
                          "PUBLISHED"
                        ]
                      },
                      "address": {
                        "type": "string"
                      },
                      "city": {
                        "type": "string"
                      },
                      "phone": {
                        "nullable": true,
                        "type": "string"
                      },
                      "website": {
                        "nullable": true,
                        "type": "string"
                      },
                      "isPartner": {
                        "type": "boolean"
                      },
                      "kwatroPerk": {
                        "nullable": true,
                        "type": "string"
                      },
                      "acceptsUnaccompaniedMinors": {
                        "type": "boolean"
                      },
                      "openingHours": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "weekday": {
                              "type": "integer",
                              "minimum": -9007199254740991,
                              "maximum": 9007199254740991
                            },
                            "opensAtMinute": {
                              "type": "integer",
                              "minimum": -9007199254740991,
                              "maximum": 9007199254740991
                            },
                            "closesAtMinute": {
                              "type": "integer",
                              "minimum": -9007199254740991,
                              "maximum": 9007199254740991
                            }
                          },
                          "required": [
                            "weekday",
                            "opensAtMinute",
                            "closesAtMinute"
                          ]
                        }
                      },
                      "createdAt": {
                        "type": "string",
                        "format": "date-time",
                        "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                      }
                    },
                    "required": [
                      "id",
                      "slug",
                      "name",
                      "type",
                      "status",
                      "address",
                      "city",
                      "phone",
                      "website",
                      "isPartner",
                      "kwatroPerk",
                      "acceptsUnaccompaniedMinors",
                      "openingHours",
                      "createdAt"
                    ]
                  }
                }
              }
            }
          }
        },
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/venues/{id}": {
      "patch": {
        "operationId": "AdminCatalogController_updateVenue",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "PENDING",
                      "PUBLISHED"
                    ]
                  },
                  "isPartner": {
                    "type": "boolean"
                  },
                  "kwatroPerk": {
                    "nullable": true,
                    "type": "string",
                    "maxLength": 120
                  },
                  "acceptsUnaccompaniedMinors": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "slug": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "type": {
                      "type": "string",
                      "enum": [
                        "GAME_BAR",
                        "TCG_SHOP",
                        "LUDOTHEQUE",
                        "ASSOCIATION",
                        "OTHER"
                      ]
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "PENDING",
                        "PUBLISHED"
                      ]
                    },
                    "address": {
                      "type": "string"
                    },
                    "city": {
                      "type": "string"
                    },
                    "phone": {
                      "nullable": true,
                      "type": "string"
                    },
                    "website": {
                      "nullable": true,
                      "type": "string"
                    },
                    "isPartner": {
                      "type": "boolean"
                    },
                    "kwatroPerk": {
                      "nullable": true,
                      "type": "string"
                    },
                    "acceptsUnaccompaniedMinors": {
                      "type": "boolean"
                    },
                    "openingHours": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "weekday": {
                            "type": "integer",
                            "minimum": -9007199254740991,
                            "maximum": 9007199254740991
                          },
                          "opensAtMinute": {
                            "type": "integer",
                            "minimum": -9007199254740991,
                            "maximum": 9007199254740991
                          },
                          "closesAtMinute": {
                            "type": "integer",
                            "minimum": -9007199254740991,
                            "maximum": 9007199254740991
                          }
                        },
                        "required": [
                          "weekday",
                          "opensAtMinute",
                          "closesAtMinute"
                        ]
                      }
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                    }
                  },
                  "required": [
                    "id",
                    "slug",
                    "name",
                    "type",
                    "status",
                    "address",
                    "city",
                    "phone",
                    "website",
                    "isPartner",
                    "kwatroPerk",
                    "acceptsUnaccompaniedMinors",
                    "openingHours",
                    "createdAt"
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "admin"
        ]
      }
    },
...
    "/admin/events/{id}/cancel": {
      "post": {
        "operationId": "AdminCatalogController_cancelEvent",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reason": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 500
                  },
                  "series": {
                    "default": false,
                    "type": "boolean"
                  }
                },
                "required": [
                  "reason"
                ]
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "admin"
        ]
      }
    },
...
    "/admin/games/{id}/merge": {
      "post": {
        "operationId": "AdminCatalogController_merge",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "intoId": {
                    "type": "string",
                    "minLength": 1
                  }
                },
                "required": [
                  "intoId"
                ]
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "admin"
        ]
      }
    }
  },
  "info": {
    "title": "API Kwatro",
    "description": "",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "dev-user": {
        "type": "apiKey",
        "in": "header",
        "name": "x-dev-user-id"
      }
    },
...
  },
  "security": [
    {
      "dev-user": []
    }
  ]
}